WIP: assign training

+++ perseu/lib/src/models/requests/assign_training.dart
@@ -4,7 +4,7 @@ import 'package:perseu/src/models/training_model.dart';
 
 class AssignTrainingRequest {
   final TrainingModel training;
-  final List<double> athletesIds;
+  final List<int> athletesIds;
   final int coachId;
 
   AssignTrainingRequest(this.athletesIds, this.training, this.coachId);
